agent/requirements.txt
google-adk[eval]
pydantic
google-genai
opentelemetry-instrumentation-google-genai
